Integrate GitHub assignment polling into message handler
