+++ app/main_message_handler.py
@@ -22,6 +22,16 @@ from app.broker import (
     TASK_QUEUE_NAME,
     TaskQueueMessage,
 )
+from app.github_ingress_runtime import (
+    AssignmentCheckpoint,
+    GitHubAssignmentCheckpointStore,
+    checkpoint_scope_key,
+    default_graphql_runner,
+    fetch_assignment_events_for_repository,
+    parse_repo_slug,
+    publish_assignment_events,
+    select_events_after_checkpoint,
+)
 from app.main import (
     TELEGRAM_MEMORY_TURN_LIMIT,
     _build_email_sender,
@@ -66,9 +76,29 @@ _ALLOWED_CONFIG_KEYS = frozenset(
         "telegram_allowed_chat_ids",
         "telegram_allowed_channel_ids",
         "telegram_context_refs",
+        "github_repositories",
+        "github_assignee_login",
+        "github_reply_channel_type",
+        "github_reply_channel_target",
+        "github_context_refs",
+        "github_policy_profile",
+        "github_max_issues",
+        "github_max_timeline_events",
     }
 )
 BBMB_EGRESS_PICKUP_WAIT_SECONDS = 5
+
+
+@dataclass(frozen=True)
+class GitHubIngressSettings:
+    repositories: list[str]
+    assignee_login: str
+    reply_channel_type: str
+    reply_channel_target: str
+    context_refs: list[str]
+    policy_profile: str
+    max_issues: int
+    max_timeline_events: int
 
 
 @dataclass
@@ -218,6 +248,28 @@ def _parse_args() -> argparse.Namespace:
         default=[],
         help="Generic context ref (repeatable).",
     )
+    parser.add_argument(
+        "--github-repository",
+        action="append",
+        default=[],
+        help="GitHub repository in owner/repo format (repeatable).",
+    )
+    parser.add_argument("--github-assignee-login", help="GitHub login to filter assigned events for.")
+    parser.add_argument("--github-reply-channel-type", help="Reply channel type for generated tasks.")
+    parser.add_argument("--github-reply-channel-target", help="Reply channel target for generated tasks.")
+    parser.add_argument(
+        "--github-context-ref",
+        action="append",
+        default=[],
+        help="Context ref to attach to generated tasks (repeatable).",
+    )
+    parser.add_argument("--github-policy-profile", help="Policy profile for generated tasks.")
+    parser.add_argument("--github-max-issues", type=_positive_int, help="Per-repo issue scan limit.")
+    parser.add_argument(
+        "--github-max-timeline-events",
+        type=_positive_int,
+        help="Per-issue assigned-event scan limit.",
+    )
     return parser.parse_args()
 
 
@@ -296,6 +348,159 @@ def _resolve_allowed_egress_channels(args: argparse.Namespace, config: dict[str,
     if any(not item.strip() for item in merged):
         raise ValueError("allowed_egress_channel entries must not be empty")
     return set(merged)
+
+
+def _resolve_required_str(cli_value: str | None, config_value: object, *, setting_name: str) -> str:
+    resolved = _resolve_str(
+        cli_value,
+        config_value,
+        default_value="",
+        setting_name=setting_name,
+    ).strip()
+    if not resolved:
+        raise ValueError(f"{setting_name} is required when github_repositories is configured")
+    return resolved
+
+
+def _resolve_github_repositories(args: argparse.Namespace, config: dict[str, object]) -> list[str]:
+    repositories: list[str] = []
+    config_values = config.get("github_repositories")
+    if config_values is not None:
+        if not isinstance(config_values, list) or not all(isinstance(item, str) for item in config_values):
+            raise ValueError("config github_repositories must be a list of owner/repo strings")
+        repositories.extend(config_values)
+    repositories.extend(args.github_repository)
+
+    deduped: list[str] = []
+    seen: set[str] = set()
+    for repository in repositories:
+        owner, name = parse_repo_slug(repository)
+        normalized = f"{owner}/{name}"
+        if normalized in seen:
+            continue
+        seen.add(normalized)
+        deduped.append(normalized)
+    return deduped
+
+
+def _resolve_github_context_refs(args: argparse.Namespace, config: dict[str, object]) -> list[str]:
+    context_refs: list[str] = []
+    config_values = config.get("github_context_refs")
+    if config_values is not None:
+        if not isinstance(config_values, list) or not all(isinstance(item, str) for item in config_values):
+            raise ValueError("config github_context_refs must be a list of strings")
+        context_refs.extend(config_values)
+    context_refs.extend(args.github_context_ref)
+    if any(not item.strip() for item in context_refs):
+        raise ValueError("github_context_ref entries must not be empty")
+    return [item.strip() for item in context_refs]
+
+
+def _resolve_github_ingress_settings(
+    args: argparse.Namespace,
+    config: dict[str, object],
+) -> GitHubIngressSettings | None:
+    repositories = _resolve_github_repositories(args, config)
+    if not repositories:
+        return None
+
+    return GitHubIngressSettings(
+        repositories=repositories,
+        assignee_login=_resolve_required_str(
+            args.github_assignee_login,
+            config.get("github_assignee_login"),
+            setting_name="github_assignee_login",
+        ),
+        reply_channel_type=_resolve_required_str(
+            args.github_reply_channel_type,
+            config.get("github_reply_channel_type"),
+            setting_name="github_reply_channel_type",
+        ),
+        reply_channel_target=_resolve_required_str(
+            args.github_reply_channel_target,
+            config.get("github_reply_channel_target"),
+            setting_name="github_reply_channel_target",
+        ),
+        context_refs=_resolve_github_context_refs(args, config),
+        policy_profile=_resolve_str(
+            args.github_policy_profile,
+            config.get("github_policy_profile"),
+            default_value="default",
+            setting_name="github_policy_profile",
+        ).strip(),
+        max_issues=_resolve_positive_int(
+            args.github_max_issues,
+            config.get("github_max_issues"),
+            default_value=25,
+            setting_name="github_max_issues",
+        ),
+        max_timeline_events=_resolve_positive_int(
+            args.github_max_timeline_events,
+            config.get("github_max_timeline_events"),
+            default_value=10,
+            setting_name="github_max_timeline_events",
+        ),
+    )
+
+
+def _poll_github_assignment_ingress(
+    *,
+    settings: GitHubIngressSettings,
+    checkpoint_store: GitHubAssignmentCheckpointStore,
+    store: SQLiteStateStore,
+    broker: BBMBQueueAdapter,
+) -> tuple[int, int, int, str]:
+    scope_key = checkpoint_scope_key(
+        repositories=settings.repositories,
+        assignee_login=settings.assignee_login,
+    )
+    checkpoint = checkpoint_store.get_checkpoint(scope_key)
+    events = []
+    scanned_event_count = 0
+    for repository in settings.repositories:
+        owner, name = parse_repo_slug(repository)
+        try:
+            repository_events = fetch_assignment_events_for_repository(
+                repo_owner=owner,
+                repo_name=name,
+                assignee_login=settings.assignee_login,
+                issue_limit=settings.max_issues,
+                timeline_limit=settings.max_timeline_events,
+                graphql_runner=default_graphql_runner,
+            )
+        except Exception:  # noqa: BLE001
+            LOGGER.exception(
+                "github_assignment_poll_failed repository=%s assignee=%s",
+                repository,
+                settings.assignee_login,
+            )
+            continue
+        scanned_event_count += len(repository_events)
+        events.extend(repository_events)
+
+    new_events = select_events_after_checkpoint(events, checkpoint=checkpoint)
+    published_count = publish_assignment_events(
+        events=new_events,
+        store=store,
+        broker=broker,
+        reply_channel_type=settings.reply_channel_type,
+        reply_channel_target=settings.reply_channel_target,
+        context_refs=settings.context_refs,
+        policy_profile=settings.policy_profile,
+    )
+    checkpoint_id = checkpoint.event_id if checkpoint else "none"
+    if new_events:
+        latest = new_events[-1]
+        checkpoint_store.set_checkpoint(
+            scope_key,
+            checkpoint=AssignmentCheckpoint(
+                event_created_at=latest.event_created_at,
+                event_id=latest.event_id,
+            ),
+        )
+        checkpoint_id = latest.event_id
+
+    return scanned_event_count, len(new_events), published_count, checkpoint_id
 
 
 def _build_policy_decision_for_message(egress_message: EgressQueueMessage) -> PolicyDecision:
@@ -530,12 +735,14 @@ def main() -> int:
         setting_name="poll_timeout_seconds",
     )
     allowed_egress_channels = _resolve_allowed_egress_channels(args, config)
+    github_ingress_settings = _resolve_github_ingress_settings(args, config)
 
     store = SQLiteStateStore(db_path)
     ledger = TaskLedgerStore(db_path)
     broker = BBMBQueueAdapter(address=bbmb_address)
     broker.ensure_queue(TASK_QUEUE_NAME)
     broker.ensure_queue(EGRESS_QUEUE_NAME)
+    github_checkpoint_store = GitHubAssignmentCheckpointStore(db_path) if github_ingress_settings else None
 
     connectors = _build_live_connectors(args, config)
     applier = IntegratedApplier(
@@ -569,6 +776,23 @@ def main() -> int:
                 store.mark_seen(envelope.source, envelope.dedupe_key)
                 ingress_published += 1
 
+        github_scanned_events = 0
+        github_new_events = 0
+        github_published = 0
+        github_checkpoint = "disabled"
+        if github_ingress_settings is not None and github_checkpoint_store is not None:
+            (
+                github_scanned_events,
+                github_new_events,
+                github_published,
+                github_checkpoint,
+            ) = _poll_github_assignment_ingress(
+                settings=github_ingress_settings,
+                checkpoint_store=github_checkpoint_store,
+                store=store,
+                broker=broker,
+            )
+
         egress_picked = broker.pickup_json(
             EGRESS_QUEUE_NAME,
             timeout_seconds=poll_timeout_seconds,
@@ -590,6 +814,7 @@ def main() -> int:
         LOGGER.info(
             (
                 "message_handler_loop_completed loop=%s ingress_published=%s "
+                "github_scanned_events=%s github_new_events=%s github_published=%s github_checkpoint=%s "
                 "egress_received_total=%s egress_dispatched_total=%s egress_deduped_total=%s "
                 "egress_dedupe_hit_rate_pct=%s egress_dropped_total=%s "
                 "egress_dispatch_latency_ms_avg=%s egress_dispatch_latency_ms_max=%s "
@@ -597,6 +822,10 @@ def main() -> int:
             ),
             loop_count,
             ingress_published,
+            github_scanned_events,
+            github_new_events,
+            github_published,
+            github_checkpoint,
             telemetry_snapshot["received_total"],
             telemetry_snapshot["dispatched_total"],
             telemetry_snapshot["deduped_total"],
