--- configs/chatting-github-ingress.env.example
@@ -1,1 +1,0 @@
-CHATTING_GITHUB_INGRESS_CONFIG_PATH=/home/edward/chatting/configs/github-ingress-runtime.local.json
