--- configs/github-ingress-runtime.example.json
@@ -1,18 +1,0 @@
-{
-  "bbmb_address": "127.0.0.1:9876",
-  "db_path": "/home/edward/chatting/state/github-ingress.db",
-  "poll_interval_seconds": 60,
-  "github_repositories": [
-    "brokensbone/chatting"
-  ],
-  "github_assignee_login": "BillyAcachofa",
-  "github_reply_channel_type": "telegram",
-  "github_reply_channel_target": "8605042448",
-  "github_context_refs": [
-    "repo:/home/edward/chatting",
-    "repo:/home/edward/bbmb"
-  ],
-  "github_policy_profile": "default",
-  "github_max_issues": 25,
-  "github_max_timeline_events": 10
-}
