--- deploy/systemd/chatting-github-ingress.service
@@ -1,19 +1,0 @@
-[Unit]
-Description=Chatting GitHub Assignment Ingress Service
-After=network-online.target
-Wants=network-online.target
-
-[Service]
-Type=simple
-User=edward
-Group=edward
-WorkingDirectory=/home/edward/chatting
-EnvironmentFile=/home/edward/chatting/configs/chatting-github-ingress.env
-ExecStart=/usr/bin/python3 -m app.main_github_ingress
-Restart=always
-RestartSec=5
-TimeoutStopSec=20
-SyslogIdentifier=chatting-github-ingress
-
-[Install]
-WantedBy=multi-user.target
